move traits structural impls to traits::structural_impls
